requirements.txt
-e .
colour-runner==0.0.4
coverage==3.7.1
django-admin-sso==0.1.3
django-crispy-forms==1.4.0
django-discover-runner==1.0
dj-database-url==0.3.0
factory-boy==2.4.1
flake8==2.2.5
flake8-import-order==0.5.1
incuna-test-utils==6.0.0
mock==1.0.1
psycopg2==2.5.4
